rtosc.c: Fixed float spec conformance
- Floating point numbers now match the spec's endianess
- Test added to compare the output of rtosc_message() for the OSC 1.0 Spec
  examples

+++ include/rtosc/rtosc.h
@@ -41,7 +41,7 @@ typedef struct {
 typedef union {
     int32_t i;
     char T;
-    double f;
+    float f;
     const char *s;
     blob_t b;
 } arg_t;
